Better support for incrementals on Solr cloud
This reduces the problems encountered when using incremental backups
on Solr cloud. When shard leaders change, it's possible for segment files
to be the same size but have different content, leading to incrementals
taken across multiple replicas being corrupt.

Adds a new option `--consistant` - which selects a shard replica based on
name ordering.